+++ requake/map_families.py
@@ -24,7 +24,7 @@ from .map_tiles import (
     EsriImagery,
     StamenTerrain,
 )
-from .families import read_selected_families
+from .families import FamilyNotFoundError, read_selected_families
 from .rq_setup import rq_exit
 logger = logging.getLogger(__name__.rsplit('.', maxsplit=1)[-1])
 # Reduce logging level for Matplotlib to avoid DEBUG messages
@@ -100,7 +100,7 @@ def _make_basemap(config):
 def map_families(config):
     try:
         families = read_selected_families(config)
-    except Exception as msg:
+    except FamilyNotFoundError as msg:
         logger.error(msg)
         rq_exit(1)
     fig, ax = _make_basemap(config)